Do not specific tooling if the docker compose does not exists

+++ helpers.sh
@@ -34,7 +34,7 @@ install_tooling() {
 
         [ -e /usr/local/bin/$scriptname ] && rm /usr/local/bin/$scriptname
         if [ -e "$scriptname" ]; then
-            if [ "$dependency" == "*" ] || grep -q "$dependency" "$BTCPAY_DOCKER_COMPOSE"; then
+            if [ "$dependency" == "*" ] || ( [ -e "$BTCPAY_DOCKER_COMPOSE" ] && grep -q "$dependency" "$BTCPAY_DOCKER_COMPOSE" ); then
                 chmod +x $scriptname
                 ln -s "$(pwd)/$scriptname" /usr/local/bin
                 echo "Installed $scriptname to /usr/local/bin: $comment"
